torchlego/requirements.txt
Pillow
pydantic
pyyaml
sanic
torch
torchvision